Move pickers into components/Picker and move ColorInput into components"
