+++ src/components/ColorInput.css
