+++ src/components/ColorInput.tsx
@@ -1,11 +1,11 @@
 import React, { useCallback } from 'react';
 
-import { useColorSpace } from '../../hooks/useColorSpace';
-import { Color, InputValue } from '../../types';
-import { clampTo0_1, roundArrayTo1Decimals, roundTo1Decimals } from '../../util/mathUtils';
-import { inputValueToNumber, inputValueToString } from '../../util/parsingUtils';
-import { hex_to_rgb, rgb_to_hex } from '../../color/general';
-import Input from '../Lib/Input';
+import { useColorSpace } from '../hooks/useColorSpace';
+import { Color, InputValue } from '../types';
+import { clampTo0_1, roundArrayTo1Decimals, roundTo1Decimals } from '../util/mathUtils';
+import { inputValueToNumber, inputValueToString } from '../util/parsingUtils';
+import { hex_to_rgb, rgb_to_hex } from '../color/general';
+import Input from './Lib/Input';
 
 import './ColorInput.css';
