+++ src/components/Picker/SliderPicker.css
